remove kscreen rotation hack on x11
it has moved to an xsession.d script so it runs before kscreen, allowing
kscreen to correctly set up multiple displays with the dock. the
previous hack caused the rotation to occur to late and ending up
rotating such that the virtual screens would overlap when another screen
was plugged into the dock.

the hack is still necessary for wayland. there are some options on the
table to fix the rotation in a neater fashion there.

also see
https://gitlab.steamos.cloud/jupiter/steamdeck-kde-presets/-/merge_requests/27