steamos-add-to-steam: Don't create an intermediate .desktop file
This script adds an external game to Steam by creating a temporary
.desktop file and using steam://addnonsteamgame/ to add it. There
are a few problems with this:

1. There is not need to create this file because Steam can add normal
   binaries just fine and the .desktop file does not provide any
   additional metadata.

2. The .desktop file parser in Steam is currently broken[1] so if the
   path of the executable has e.g a whitespace the resulting entry
   won't work.

3. If the binary is not executable then Steam ignores it and uses the
   temporary desktop file as if it was the game binary. This doesn't
   work and even if it did it would anyway disappear after a reboot.

This patch removes this intermediate step and adds the provided file
directly in the cases where it makes sense (.exe files, that need to
be launched with Proton). In the cases where the actual file needs to
be executable (regular Linux binaries, shell scripts, AppImage files)
it shows an error instead.

[1] https://github.com/ValveSoftware/steam-for-linux/issues/8164