nested-desktop: Use a custom XDG_RUNTIME_DIR
Sharing the XDG_RUNTIME_DIR with gamescope with a separate dbus bus
leads to two instances of the xdg-desktop-portal fighting over the same
file path. This has a range of side effects, including breaking all file
transfers.

Using a separate directory solves this, but creates a few new problems
with pulseaudio that needs manual fixup.

This also sets the path to having a nested systemd user session.