Add a separate global theme for the deck
Deck sepcific things go into the deck theme this includes:
- setting the kickoff icon to deck logo
- setting the default wallpaper to the deck one
- the splash screen with the deck logo

Before the global theme was called com.valve.vapor.desktop.
The theme without deck logos now takes this name. The deck one
becomes com.valve.vapor.deck.desktop. This should not be a problem
for the following reasons: Since it was the default theme, it's
not saved in user config, on next startup plasma detects the mismatch
between active and current global theme and updates it to the
active global theme. Both themes are basically the same settings-wise.

A small systemd service is included that updates the default theme
in /etc/xdg/kdeglobals to com.valve.vapor.deck when running on a steam
deck.