Make Return To Gaming Mode icon forcefully go to gamescope
The previous Return To Gaming mode icon performed a logout. This was
sufficient at the time because SteamOS had a hardcoded default. We now
want to cover the use case of defaulting to Plasma and being able to
enter gaming mode. The new script makes use of new steamos-manager API
to perform the switch.

As the file may already exist on the user's desktop, it is synced from
the base system on startup.

The symlink line is changed so we don't bail the script out early if the
link exists.