steamos-add-to-steam: show help if a file is not specified
This script assumes that the user always provides a file name, and if
one is not set then it fails with a confusing error message:

   realpath: '': No such file or directory

This case is handled with this patch.

In addition to that, realpath is also run with '--' to ensure that the
file name is not misinterpreted as a command-line option.